+++ NEWS.md
@@ -1,5 +1,8 @@
 # roxygen2 (development version)
 
+* `@export` will now export both the class and constructor function when
+  applied to expressions like `foo <- setClass("foo")` (#1216).
+
 * Generated HTML for code blocks never includes "NA" for language (#1251). 
 
 * Using a level 1 heading in the wrong tag now gives a more useful warning